requirements.txt
apscheduler>=3.10,<4
pymupdf>=1.23
python-docx>=1.1
openpyxl>=3.1
pytesseract>=0.3
Pillow>=10.0
pyyaml>=6.0
requests>=2.31
